fix(angular): make package version detection more robust
The command used to determine the latest AngularJS package version was
assuming that the return JSON response was pretty-formatted (with one
key-value pair per line). However, in recent versions of the registry,
the JSON response is formatted without whitespace to reduce the response
size.

Make the command more robust by ensuring it can handle the returned JSON
response regardless of its formatting.